+++ components/Layout/index.tsx
@@ -13,7 +13,6 @@
  * @module
  */
 import setup, { iLayout } from './setup.ts';
-import { ComponentChild } from 'preact';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /**
@@ -21,10 +20,7 @@ import { ComponentChild } from 'preact';
  *
  * [Component properties are partial](https://deno.land/x/lunchbox#configure-anything-easily)
  *
- * @param {LayoutTypes} type
- *    Available options that represent the different column arrangements on the grid.
- *
- * @param {boolean} whitespaceMode
+ * @param {boolean} whitespace
  *    If true, adds horizontal whitespace margins to the layout section. This is useful for making
  *    certain layout types more focused. Additionally, having this mode turned off would create a
  *    layout with the minimum whitespace, making it have a "dashboard-like" appearance.
@@ -33,41 +29,11 @@ import { ComponentChild } from 'preact';
  *  The `<Layout />` component.
  */
 export default function Layout(props: Partial<iLayout>) {
-  const {
-    c,
-    nostyle,
-    nostyleAll,
-    fref,
-    fwd,
-    type,
-    whitespaceMode,
-    children,
-    ...p
-  } = setup(
-    props,
-  );
+  const { whitespace, children, ...p } = setup(props);
 
   return (
-    <div ref={fref} {...p} class={c.layout}>
-      {Array.isArray(children)
-        ? (
-          children.map((child: ComponentChild) =>
-            child
-              ? (
-                <div ref={fwd.module?.ref} class={c.module} {...fwd.module}>
-                  {child}
-                </div>
-              )
-              : null
-          )
-        )
-        : children
-        ? (
-          <div ref={fwd.module?.ref} class={c.module} {...fwd.module}>
-            {children}
-          </div>
-        )
-        : null}
-    </div>
+    <section {...p}>
+      {children}
+    </section>
   );
 }