+++ components/Layout/setup.ts
@@ -9,26 +9,19 @@
  *
  * @module
  */
-import { apDef, o, part } from '../../src/utils.ts';
-import { iComponent, iFwd } from '../../src/types.ts';
-import { LayoutTypes } from '../../src/enums.ts';
+import { apDef, o } from '../../src/utils.ts';
+import { iComponent } from '../../src/types.ts';
 import { styles } from './styles.ts';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** Properties of the `<Layout />` component. */
-export type iLayout = iComponent<HTMLDivElement> & {
-  type: LayoutTypes;
-  whitespaceMode: boolean;
-  fwd: Partial<{
-    module: iFwd<HTMLDivElement>;
-  }>;
+export type iLayout = iComponent & {
+  whitespace: boolean;
 };
 
 /** Default values of the `<Layout />` component's props. */
 const defaults: iLayout = {
-  whitespaceMode: false,
-  type: 'default',
-  fwd: {},
+  whitespace: false,
 };
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -36,23 +29,10 @@ const defaults: iLayout = {
 export default (props: Partial<iLayout>) => {
   const p = apDef<iLayout>(defaults, props);
 
-  const { module } = p.fwd;
+  p.class = o(
+    [styles, `layout${p.whitespace ? '--whitespace' : ''}`],
+    { ...p },
+  );
 
-  const classes = part({
-    layout: o(
-      [
-        styles,
-        props.type === 'default'
-          ? null
-          : props.whitespaceMode
-          ? 'layout--whitespace'
-          : 'layout',
-      ],
-      { ...p },
-    ),
-    module: o(`layout__module--${props.type}`, { ...module }),
-  });
-
-  delete p.class;
-  return { c: classes, ...p };
+  return p;
 };