+++ src/types.ts
@@ -10,6 +10,7 @@
  * @module
  */
 import { JSX } from 'preact';
+import { SignalLike } from '$fresh/src/types.ts';
 
 /**
  * This type is used for standarizing all components. By design every component must have a protagonic
@@ -57,3 +58,6 @@ export type iFwd<T extends EventTarget = HTMLElement> = iElement<T> & {
  * `deno-lint(ban-types) recommends as the correct way to express the type of an empty object.`.
  */
 export type EmptyObject = Record<string | number | symbol, never>;
+
+/** */
+export type ClassNameString = SignalLike<string | undefined> | string;